Refactor operator routing
